My attempt at directory structure (WIP)
