+++ src/index.js
@@ -1,14 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import App from "./App";
+import CounterList from "./CounterList";
 
 //import Counter from "./Counter";
 
 const rootElement = document.getElementById("root2");
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  rootElement
-);
+ReactDOM.render(<CounterList />, rootElement);
